+++ plone/app/vocabularies/utils.py
@@ -1,20 +1,7 @@
-from plone.app.querystring import queryparser
-from plone.app.querystring.interfaces import IParsedQueryIndexModifier
-from zope.component import getUtilitiesFor
+from zope.deferredimport import deprecated
 
 
-def parseQueryString(context, query):
-    parsedquery = queryparser.parseFormquery(context, query)
-
-    index_modifiers = getUtilitiesFor(IParsedQueryIndexModifier)
-    for name, modifier in index_modifiers:
-        if name in parsedquery:
-            new_name, query = modifier(parsedquery[name])
-            parsedquery[name] = query
-            # if a new index name has been returned, we need to replace
-            # the native ones
-            if name != new_name:
-                del parsedquery[name]
-                parsedquery[new_name] = query
-
-    return parsedquery
+deprecated(
+    "Import parseQueryString as parseAndModifyFormquery from plone.app.querystring.queryparser instead (will be removed in Plone 7)",
+    parseQueryString="plone.app.querystring.queryparser:parseAndModifyFormquery",
+)