+++ CHANGES.rst
@@ -8,6 +8,39 @@ Changelog
 
 .. towncrier release notes start
 
+5.0.3 (2023-05-22)
+------------------
+
+Bug fixes:
+
+
+- Fix a circular dependency to `plone.app.querystring`.
+  Move `.catalog.CatalogVocabularyFactory` to `plone.app.querystring.vocabularies`, move the ZCML to register the factory, move the the test and put BBB code with deprecation wanring into place.
+  Move `.utils.parse_query` with new name `parseAndModifyFormquery` to `plone.app.querystring.queryparser` and put BBB code with deprecation wanring into place.
+  [@jensens] (fix-circular-dep-paquerstring)
+
+
+5.0.2 (2023-04-06)
+------------------
+
+Bug fixes:
+
+
+- Import navigation root specific from plone.base.
+  And so remove transitive circular dependency on plone.app.layout.
+  [jensens] (#74)
+
+
+5.0.1 (2023-03-22)
+------------------
+
+Internal:
+
+
+- Update configuration files.
+  [plone devs] (b2d5d4a5)
+
+
 5.0.0 (2022-11-30)
 ------------------
 
@@ -370,7 +403,7 @@ Fixes:
 2.1.15 (2014-04-11)
 -------------------
 
-- Make ``KeywordsVocabulary`` more customizeable using an ``keyword_index``
+- Make ``KeywordsVocabulary`` more customizable using an ``keyword_index``
   class variable to allow users to inherit and just override that attribute
   to build their own keyword vocabularies.
   [saily]